[PATCH] Fix JSON export, categorylinks migration support
JSON export: fix `.mediawiki` suffix for non-wikitext pages

In https://github.com/WikiTeq/PagePort/commit/a6b800c, the `.mediawiki` suffix
was removed from non-wikitext pages in the output files that are created.
However, when exporting the structure of the files to JSON, the URLs to the
pages were not updated, meaning that they always included `.mediawiki` and,
for non-wikitext pages, were broken.

Make the `.mediawiki` suffix in URLs conditional on the page content model
being wikitext, and add test cases to confirm.

Categorylinks: add support for reading from the new schema implemented in
T299951. For the sake of simplicity, we do not support running PagePort to
export the contents of a category while the categorylinks migration is in
progress, i.e. the read stage must be SCHEMA_COMPAT_READ_OLD or
SCHEMA_COMPAT_READ_NEW.

WE-512

Co-authored: Daniel Scherzer <daniel@wikiteq.com>

T299951

Change-Id: I53e1d7917cab39f96126ed910ae6901ec5fa3a71